requirements.txt
requests
psycopg2-binary
flask
flask_restful
